--- C++ quality questions/Segment_tree_cf.cpp
@@ -1,294 +1,0 @@
-//Question link:
-//https://codeforces.com/contest/474/problem/E
-
-#include <bits/stdc++.h>
-using namespace std;
-#define int               long long
-#define pb                push_back
-#define ppb               pop_back
-#define pf                push_front
-#define ppf               pop_front
-#define all(x)            (x).begin(),(x).end()
-#define sz(x)             (int)((x).size())
-#define fr                first
-#define sc                second
-#define pii               pair<int,int>
-#define rep(i,a,b)        for(int i=a;i<b;i++)
-#define mem1(a)           memset(a,-1,sizeof(a))
-#define mem0(a)           memset(a,0,sizeof(a))
-#define ppc               __builtin_popcount
-#define ppcll             __builtin_popcountll
-
-template<typename T1,typename T2>istream& operator>>(istream& in,pair<T1,T2> &a){in>>a.fr>>a.sc;return in;}
-template<typename T1,typename T2>ostream& operator<<(ostream& out,pair<T1,T2> a){out<<a.fr<<" "<<a.sc;return out;}
-template<typename T,typename T1>T amax(T &a,T1 b){if(b>a)a=b;return a;}
-template<typename T,typename T1>T amin(T &a,T1 b){if(b<a)a=b;return a;}
-
-const long long INF=1e18;
-const int32_t M=1e9+7;
-const int32_t MM=998244353;
-
-const int N=1e5+5;
-
-
-const int32_t maxn=N;
-template<typename NODE,typename UPDATE>
-struct segtree{
-	bool built=false,lazy[4*maxn];
-	NODE zero=NODE(),t[4*maxn];
-	UPDATE noop=UPDATE(),upds[4*maxn];
-	int32_t tl[4*maxn],tr[4*maxn];
-	inline void pushdown(int32_t v)
-	{
-		if(lazy[v]){
-			apply(v*2,upds[v]);
-			apply(v*2+1,upds[v]);
-			lazy[v]=0;
-			upds[v]=noop;
-		}
-	}
-	inline void apply(int32_t v,UPDATE &val)
-	{
-		if(tl[v]!=tr[v]){
-			lazy[v]=1;
-			upds[v].combine(val,tl[v],tr[v]);
-		}
-		val.apply(t[v],tl[v],tr[v]);
-	}
-	template<typename T>
-	void build(T a, int32_t v, int32_t l, int32_t r) {
-		tl[v]=l;
-		tr[v]=r;
-		if (l == r) {
-			t[v]=a[l];
-			return;
-		} else {
-			int32_t tm = (l + r) / 2;
-			build(a, v*2, l, tm);
-			build(a, v*2+1, tm+1, r);
-			t[v].merge(t[v*2],t[v*2+1]);
-		}
-	}
-	NODE query(int32_t v, int l, int r) {
-		if (l > tr[v] || r < tl[v])
-			return zero;
-		if (l <= tl[v] && tr[v] <= r) {
-			return t[v];
-		}
-		pushdown(v);
-		NODE a,b,ans;
-		a=query(v*2, l, r);
-		b=query(v*2+1, l, r);
-		ans.merge(a,b);
-		return ans;
-	}
-	void rupd(int32_t v, int l, int r, UPDATE &val)
-	{
-		if(l > tr[v] || r < tl[v])
-			return;
-		if(l <= tl[v] && tr[v] <= r)
-		{
-			apply(v,val);
-			return;
-		}
-		pushdown(v);
-		rupd(v*2,l,r,val);
-		rupd(v*2+1,l,r,val);
-		t[v].merge(t[v*2],t[v*2+1]);
-	}
-	template<typename T>
-	int descent_right(int l, T x, int32_t v, NODE &prev) {
-		if (l > tr[v]) // node is completely out of range
-			return len;
-		if(l <= tl[v]){ // node is completely in range
-			NODE cur;
-			cur.merge(prev,t[v]);
-			if (!cur.check(x)){ // go further right than this node
-				swap(prev,cur); // merging this node's contribution
-				return len;
-			}
-			if (tl[v]==tr[v]) {
-				return tr[v];
-			}
-		}
-		pushdown(v);
-		int ans=descent_right(l, x, v*2, prev); // trying to find in left child
-		if(ans!=len)return ans; // found in left child
-		return descent_right(l, x, v*2+1, prev); // finding in right child
-	}
-	template<typename T>
-	int descent_left(int r, T x, int32_t v, NODE &prev) {
-		if (r < tl[v]) // node is completely out of range
-			return -1;
-		if(r >= tr[v]){ // node is completely in range
-			NODE cur;
-			cur.merge(t[v],prev);
-			if (!cur.check(x)){ // go further left than this node
-				swap(cur,prev); // merging this node's contribution
-				return -1;
-			}
-			if (tl[v]==tr[v]) {
-				return tl[v];
-			}
-		}
-		pushdown(v);
-		int ans=descent_left(r, x, v*2+1, prev); // trying to find in right child
-		if(ans!=-1)return ans; // found in right child
-		return descent_left(r, x, v*2, prev); // finding in left child
-	}
-	
-	int len=maxn;
-	void clear(){
-		fill(t,t+4*len,zero);
-		fill(lazy,lazy+4*len,false);
-		fill(upds,upds+4*len,noop);
-	}
-	template<typename T>
-	void build(T a){
-		build(a,1,0,len-1);
-		built=true;
-	}
-	template<typename T>
-	int descent_right(int l, T x){ // minimum r such that [l...r].check(x) == true, returns segtree.len if not found
-		NODE prev=zero;
-		return descent_right(l,x,1,prev);
-	}
-	template<typename T>
-	int descent_left(int r, T x){ // maximum l such that [l...r].check(x) == true, returns -1 if not found
-		NODE prev=zero;
-		return descent_left(r,x,1,prev);
-	}
-	NODE query(int l,int r){
-		if(!built)build(t);
-		return query(1,l,r);
-	}
-	void rupd(int l,UPDATE val){
-		if(!built)build(t);
-		rupd(1,l,l,val);
-	}
-};
-
-#define node node1
-#define update update1
-struct node{
-	int v=0;
-	int id=0;
-	node(){}
-	node(int val){
-		v=val;
-	}
-	void merge(node &l,node &r)
-	{
-		if(l.v>r.v){
-			v=l.v;
-			id=l.id;
-		}
-		else{
-			v=r.v;
-			id=r.id;
-		}
-	}
-	bool check(int x){
-		return false;
-	}
-};
-struct update{
-	int v=0;
-	int id;
-	update(){}
-	update(int val,int i){
-		v=val;
-		id=i;
-	}
-	void combine(update &other,int32_t tl,int32_t tr){
-		
-	}
-	void apply(node &x,int32_t tl,int32_t tr){
-		x.v=v;
-		x.id=id;
-	}
-};
-segtree<node,update> t;
-#undef node
-#undef update
-
-
-map<int, int> co;
-
-int h[N];
-int ans[N];
-int pre[N];
-
-void solve(){
-	int n,d;
-	cin>>n>>d;
-	rep(i,1,n+1){
-		cin>>h[i];
-		co[h[i]]=0;
-	}
-	int cnt=1;
-	for(auto &x:co){
-		x.sc=cnt++;
-	}
-	
-	// ans[1]=1;
-	// pre[1]=0;
-	// t.rupd(co[h[1]],{1,1});
-	
-	rep(i,1,n+1){
-		int k=h[i]-d;
-		auto it=co.upper_bound(k);
-		if(it!=co.begin()){
-			it--;
-			node1 z=t.query(0,it->sc);
-			ans[i]=z.v+1;
-			pre[i]=z.id;
-		}
-		
-		k=h[i]+d;
-		it=co.lower_bound(k);
-		if(it!=co.end()){
-			node1 z=t.query(it->sc,N);
-			if(z.v+1>ans[i]){
-				ans[i]=z.v+1;
-				pre[i]=z.id;
-			}
-		}
-		
-		t.rupd(co[h[i]],{ans[i],i});
-	}
-	
-	int best=1;
-	rep(i,1,n+1){
-		if(ans[i] > ans[best]){
-			best=i;
-		}
-	}
-	
-	vector<int> v;
-	while(best){
-		v.pb(best);
-		best=pre[best];
-	}
-	cout<<sz(v)<<"\n";
-	reverse(all(v));
-	for(int x:v){
-		cout<<x<<" ";
-	}
-}
-signed main(){
-	ios_base::sync_with_stdio(false);
-	cin.tie(0);cout.tie(0);
-	//freopen("input.txt", "r", stdin);
-	//freopen("output.txt", "w", stdout);
-	#ifdef SEIVE
-		seive();
-	#endif
-	#ifdef NCR
-		init();
-	#endif
-	int t=1;
-	//cin>>t;
-	while(t--) solve();
-	return 0;
-}
