+++ C++ quality questions/kruskal.cpp
