+++ C++ quality questions/prims.cpp
